Consolidate join and recursion handling
