+++ crudcrate-derive/src/codegen/join_strategies/mod.rs
@@ -1,5 +1,4 @@
-pub mod recursion;
 mod structs;
 
-pub(crate) use recursion::get_join_config;
+pub(crate) use structs::get_join_config;
 pub use structs::JoinConfig;
